// src/components/units/board/10-write/BoardWrite.types.ts
import type { ChangeEvent, MouseEvent } from "react"

export interface IBoardWriteProps {
  isEdit: boolean
  data?: any
}

export interface IMyvariables {
  number: number
  title?: string
  writer?: string
  contents?: string
}

export interface IBoardWriteUIProps {
  onClickSubmit: (event: MouseEvent<HTMLButtonElement>) => void
  onClickUpdate: (event: MouseEvent<HTMLButtonElement>) => void
  onChangeWriter: (event: ChangeEvent<HTMLInputElement>) => void
  onChangeTitle: (event: ChangeEvent<HTMLInputElement>) => void
  onChangeContents: (event: ChangeEvent<HTMLInputElement>) => void
  isEdit: boolean
  data?: any
}